requirements.txt
numpy
opencv_python
scikit_learn
flask
Werkzeug
gevent
joblib
pillow
h5py
tensorflow-cpu
gunicorn
keras
pytest
Keras-Preprocessing